% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{ca.get_empirical_quantiles}
\alias{ca.get_empirical_quantiles}
\title{Estimate empirical quantiles from predicted quantiles.}
\usage{
ca.get_empirical_quantiles(target_quantiles, predicted_quantiles, y)
}
\arguments{
\item{target_quantiles}{array of target quantiles, e.g. the cdc quantiles}

\item{predicted_quantiles}{matrix of predicted quantiles, where each row
corresponds to a location, and each column a quantiles (same scale as y)}

\item{y}{the true values on the same scale as predicted_quantiles}
}
\description{
Compute the proportion of cases below each nominal quantile. This
determines the proper quantile (after interpolation) to request in
order to achieve the target quantiles. We currently fix the median
to remain the same.
}
