% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{ds.set_default}
\alias{ds.set_default}
\title{Sets a default value if not already set}
\usage{
ds.set_default(modeling_options, key, value)
}
\arguments{
\item{modeling_options}{named list}

\item{key}{option to be set}

\item{value}{option setting}
}
\value{
a named list
}
\description{
Sets a default value if not already set
}
