% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{ds.set_default_list}
\alias{ds.set_default_list}
\title{Sets default values from a named list}
\usage{
ds.set_default_list(modeling_options, options_list)
}
\arguments{
\item{modeling_options}{named list}

\item{options_list}{named list of defaults}
}
\value{
a named list
}
\description{
Sets default values from a named list
}
