% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{ds.set_modeling_defaults}
\alias{ds.set_modeling_defaults}
\title{Sets modeling defaults}
\usage{
ds.set_modeling_defaults(modeling_options)
}
\arguments{
\item{modeling_options}{named list}
}
\value{
a named list
}
\description{
Sets modeling defaults
}
