% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecasters.R
\name{get_forecasters}
\alias{get_forecasters}
\title{Get the list of forecasters and associated forecaster types
provided by this package.}
\usage{
get_forecasters(response, geo_type, n_locations = NULL)
}
\arguments{
\item{response}{the response (e.g. "usa-facts_deaths_incidence_num")}

\item{geo_type}{the geographic type (e.g "county" or "state" or
"hrr" or "msa"... but for now only the first two),}

\item{n_locations}{the maximum number of locations to forecast, ordered by response value
descending.  Forecasts all locations when NULL.}

\item{incidence_period}{the incidence period (e.g. "epiweek" for
now, for all forecasters)}

\item{ahead}{the ahead parameter (e.g. 1, 2, 3, 4)}

\item{forecast_date}{the date of the forecast}
}
\value{
a named list, with each element of the list consisting of a
    forecaster function and type (one of `c("standalone",
    "ensemble")`). Unavailable forecasters are marked as
    `list(forecaster = NA, type = "standalone")`.
}
\description{
The evaluator will first call this function with the
    parameters shown below to determine all the forecasters
    available. It expects to get back a named list of lists of
    forecasting functions and types. If a forecasting function is
    not available for a given set of parameters, an `NA` should
    returned _instead of_ a function. This tells the evaluator to
    ignore that forecaster in a run: it ignores anything that is
    not a function. See examples in code below.
}
