% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.load_location_info}
\alias{io.load_location_info}
\title{Load health data}
\usage{
io.load_location_info(geo_type)
}
\arguments{
\item{geo_type}{the geo type}
}
\value{
tibble
}
\description{
Load health data
}
