% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{ml.fit_model}
\alias{ml.fit_model}
\title{Fit a learner to the training data}
\usage{
ml.fit_model(train_test, modeling_options)
}
\arguments{
\item{train_test}{list of training and testing data}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\description{
Fit a learner to the training data
}
