% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{ml.get_clusters}
\alias{ml.get_clusters}
\title{Finds the stratified clusters.}
\usage{
ml.get_clusters(test_X, modeling_options)
}
\arguments{
\item{test_X}{covariates as of forecast_date}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\description{
Finds the stratified clusters.
}
