% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{ml.linear}
\alias{ml.linear}
\title{Fit a simple linear model.}
\usage{
ml.linear(train_test, modeling_options)
}
\arguments{
\item{train_test}{list of training and testing data}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\description{
The following model is fit:
   response ~ sum(lag(c) for c in covariates) + log(pop) +
           sum(slope(c) for c in covariates).
The slope is computed using least squares over past lags of each signal,
for each location individually.
Calibration is not currently performed.
}
