% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{ml.stratified_linear}
\alias{ml.stratified_linear}
\title{Fit a simple linear model within stratified clusters.}
\usage{
ml.stratified_linear(train_test, modeling_options)
}
\arguments{
\item{train_test}{list of training and testing data}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\description{
Clusters are determined using a simple scheme: order the locations
by the number of cases. Sort, then divide into k equal sized groups. 
Within each cluster, the following model is fit:
   cases ~ cases_lag1 + fb_lag1 + combined_ind_lag1 + log(pop) +
           slope(cases) + slope(fb) + slope(combined_ind).
The slope is computed using least squares over past lags of each signal,
for each location individually. If in modeling_options, interactions 
are specified, all cases_lag1 interactions will be added.
Calibration is not currently performed.
}
