% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_structure.R
\name{ms.covariate}
\alias{ms.covariate}
\title{Pseudo named tuple for covariates}
\usage{
ms.covariate(
  name,
  tr = tr.identity,
  lags = c(0),
  do_rollsum = F,
  do_rollavg = F,
  do_penalize = T
)
}
\arguments{
\item{name}{name of covariate (must have match in giant rdata)}

\item{tr}{transformation function to use}

\item{lags}{the lags to use (0 for none)}

\item{do_rollsum}{bool to roll sum the covariate}

\item{do_rollavg}{bool to roll avg the covariate}

\item{do_penalize}{bool to penalize the covariate in the lasso}
}
\value{
a named list
}
\description{
Pseudo named tuple for covariates
}
