% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_structure.R
\name{ms.validate_options}
\alias{ms.validate_options}
\title{Validate the modeling options, fixing invalid values where possible.}
\usage{
ms.validate_options(modeling_options, base_df)
}
\arguments{
\item{modeling_options}{named list of modeling options}

\item{base_df}{signal data with column `time_value`}
}
\value{
a named list
}
\description{
Validate the modeling options, fixing invalid values where possible.
}
