% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pp.add_lagged_columns}
\alias{pp.add_lagged_columns}
\title{Add lagged columns to base_df. We make all possible lags,
and then select the ones we will use afterwards in pp.get_training_set.}
\usage{
pp.add_lagged_columns(base_df, modeling_options)
}
\arguments{
\item{base_df}{long data frame containing all potential data}

\item{modeling_options}{list of modeling options specified}
}
\value{
same as base_df but with additional lagged columns
}
\description{
Add lagged columns to base_df. We make all possible lags,
and then select the ones we will use afterwards in pp.get_training_set.
}
