% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pp.add_pc}
\alias{pp.add_pc}
\title{Add principal components to location_info_df}
\usage{
pp.add_pc(location_info_df, modeling_options)
}
\arguments{
\item{location_info_df}{the data frame}

\item{modeling_options}{the modeling options list}
}
\value{
tibble with columns location, population, PC01, ..., PC0n
}
\description{
Add principal components to location_info_df
}
