% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pp.get_training_set}
\alias{pp.get_training_set}
\title{Make predictors X and target y for a single forecast_date}
\usage{
pp.get_training_set(lagged_df, location_info_df, modeling_options)
}
\arguments{
\item{lagged_df}{original base_df (tall_df) with lagged columns}

\item{location_info_df}{formerly health_rankings_df}

\item{modeling_options}{the modeling options list}
}
\value{
list with names X, y, row_locations.  y is NA if modeling_options$forecast_date is in 
  the future
}
\description{
Make predictors X and target y for a single forecast_date
}
