% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pp.impute_and_select}
\alias{pp.impute_and_select}
\title{Mean impute and select relevant variables}
\usage{
pp.impute_and_select(train_test, modeling_options)
}
\arguments{
\item{train_test}{list of training and testing data}

\item{modeling_options}{list of modeling options}
}
\value{
list with names X, y, row_locations. y is NA if modeling_options$forecast_date is in the 
  future
}
\description{
Mean impute and select relevant variables
}
