% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pp.make_train_test}
\alias{pp.make_train_test}
\title{Make train predictors, targets and test predictors}
\usage{
pp.make_train_test(base_df, location_info_df, modeling_options)
}
\arguments{
\item{base_df}{the data frame}

\item{location_info_df}{formerly health_rankings_df}

\item{modeling_options}{the modeling options list}
}
\value{
list with names train_X, train_y, test_X
}
\description{
Make train predictors, targets and test predictors
}
