% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pp.select_dates}
\alias{pp.select_dates}
\title{Helper function to get for each variable the appropriate issue date and reference date.}
\usage{
pp.select_dates(base_df, forecast_date)
}
\arguments{
\item{base_df}{the data frame}

\item{forecast_date}{the forecast date}
}
\value{
a dataframe specifying the reference dates and issue dates to use for each
}
\description{
Helper function to get for each variable the appropriate issue date and reference date.
}
