% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pp.transform_and_scale}
\alias{pp.transform_and_scale}
\title{Perform additional preprocessing steps}
\usage{
pp.transform_and_scale(train_test, modeling_options)
}
\arguments{
\item{train_test}{list of training and testing data}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\value{
same as train_test, but with additional preprocessing sets and
        validation splits
}
\description{
Perform additional preprocessing steps
}
