% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{raw_forecaster}
\alias{raw_forecaster}
\title{Perform stacked forecasting}
\usage{
raw_forecaster(base_df, forecast_date, n_locations, modeling_options)
}
\arguments{
\item{base_df}{dataframe containing some covariate and response information}

\item{forecast_date}{date on which we start producing forecasts}

\item{n_locations}{the maximum number of locations to forecast, ordered by response value
descending.  Forecasts all locations when NULL.}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\value{
function that performs forecasting with proper options set
}
\description{
Perform stacked forecasting
}
