% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stacked_forecaster}
\alias{stacked_forecaster}
\title{Wrapper function to create a forecasting function}
\usage{
stacked_forecaster(
  forecast_date,
  n_locations = 200,
  modeling_options = list(learner = "stratified_linear", weeks_back = 4)
)
}
\arguments{
\item{forecast_date}{the date of the forecast}

\item{n_locations}{the number of locations (for now we will use 200
for this)}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\value{
function that performs forecasting with proper options set
}
\description{
Wrapper function to create a forecasting function
}
