% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n.R
\name{tn.get_top_n_locations}
\alias{tn.get_top_n_locations}
\title{This function returns the n** locations which have observed the largest
values of the response, cumulatively over all dates.}
\usage{
tn.get_top_n_locations(df, response, n)
}
\arguments{
\item{df}{data frame with columns geo_value, time_value,
variable_name, and value (and possibly others, which we will ignore)}

\item{response}{string indicating the name of the response}

\item{n}{how many locations should we pick? If it is greater than the total
number of locations, than we pick all of them.}
}
\description{
**: (if there are multiple locations with value equal to the nth location,
they are all included)
}
