% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{tr.column_impute}
\alias{tr.column_impute}
\title{Column mean imputation}
\usage{
tr.column_impute(matx, impute_columns = 1:ncol(matx))
}
\arguments{
\item{matx}{the matrix}

\item{impute_columns}{the columns to impute, default all}
}
\description{
Performs mean imputation on a matrix.  If all NAs, impute to zero.
}
