% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{tr.cube_root}
\alias{tr.cube_root}
\title{Cube root transformation}
\usage{
tr.cube_root(x)
}
\arguments{
\item{x}{the vector to be cube root transformed}
}
\value{
transformed vector
}
\description{
Cube root transformation
}
