% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{tr.identity}
\alias{tr.identity}
\title{Identity transformation}
\usage{
tr.identity(x)
}
\arguments{
\item{x}{the vector to be transformed}
}
\value{
transformed vector
}
\description{
Identity transformation
}
