% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{tr.inv_log_pad}
\alias{tr.inv_log_pad}
\title{Inverse padded log transformation for count data}
\usage{
tr.inv_log_pad(x, a = 1)
}
\arguments{
\item{x}{the vector to be inversely log_pad transformed}

\item{a}{the scalar constant to subtract from the vector, elementwise}
}
\value{
inversely transformed vector
}
\description{
Inverse padded log transformation for count data
}
