% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{tr.log_pad}
\alias{tr.log_pad}
\title{Padded log transformation for count data}
\usage{
tr.log_pad(x, a = 1, clip_left_0 = TRUE)
}
\arguments{
\item{x}{the vector to be log transformed}

\item{a}{the scalar constant to pad the vector, elementwise}

\item{clip_left_0}{boolean to clip negative values to 0}
}
\value{
transformed vector
}
\description{
Padded log transformation for count data
}
