% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zyzzyva-package.R
\docType{package}
\name{zyzzyva-package}
\alias{zyzzyva-package}
\title{zyzzyva is a forecaster package}
\description{
A forecaster package
}
\author{
Rob Tibshirani, Maria Jahja, Addison Hu
Maintainers: Maria Jahja<maria@stat.cmu.edu> Addison Hu<addison@stat.cmu.edu>
}
