% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{ds.set_modeling_defaults}
\alias{ds.set_modeling_defaults}
\title{Sets modeling defaults}
\usage{
ds.set_modeling_defaults(modeling_options)
}
\arguments{
\item{modeling_options}{named list}
}
\value{
a named list
}
\description{
This function takes a named list, containing user-specified modeling
options, and returns a named list, where specific modeling options that
were not explicitly set by the user are set to package defaults, as
defined in this file.  This function also verifies that only valid modeling
options (both keys and values) are passed in by the user, and if an invalid
modeling option is specified, an error will be raised.
}
