% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecasters.R
\name{get_forecasters}
\alias{get_forecasters}
\title{Get an evalcast-compliant zyzzyva forecaster based on the provided arguments.}
\usage{
get_forecasters(
  debug_folder = NULL,
  impute_last_3_responses = TRUE,
  learner = "linear",
  location_covariates = c("population"),
  log_response = TRUE,
  n_locations = NULL,
  quantiles = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  roll_lags = 7,
  seed = 2020,
  weeks_back = Inf
)
}
\arguments{
\item{debug_folder}{file path to which to write debug information.  When NULL, no debug
information is written.}

\item{impute_last_3_responses}{whether to use an ets model to impute the most recent 3 days of
the response variable.}

\item{learner}{name of learner to use}

\item{location_covariates}{list of names location covariates to use in model}

\item{log_response}{whether to log-pad the response}

\item{n_locations}{the maximum number of locations to forecast, ordered by response value
descending.  Forecasts all locations when NULL.}

\item{quantiles}{list of quantile values at which to forecast}

\item{roll_lags}{lag time in days for applying rolling sums of covariates}

\item{seed}{seed for randomness}

\item{weeks_back}{number of weeks back of data to use in training.  Automatically adjusted to
prevent avoid days not present in the data.}
}
\value{
a named list, with each element of the list consisting of a
    forecaster function and type (one of `c("standalone",
    "ensemble")`).
}
\description{
Get an evalcast-compliant zyzzyva forecaster based on the provided arguments.
}
