% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.load_location_info}
\alias{io.load_location_info}
\title{Load location information}
\usage{
io.load_location_info(geo_type, location_covariates)
}
\arguments{
\item{geo_type}{the geo type}

\item{location_covariates}{list of location-based covariates to be loaded from file}
}
\value{
tibble containing the location information
}
\description{
Location information (static in time) are provided through the "County
Health Rankings" datasets.  Therefore, legacy code may use this older
terminology.  From a modeling standpoint, any static covariates that
are location specific fall under the broader "location information"
umbrella, hence the renaming in this package.
}
