% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_forecaster.R
\name{raw_forecaster}
\alias{raw_forecaster}
\title{Perform stacked forecasting}
\usage{
raw_forecaster(base_df, modeling_options)
}
\arguments{
\item{base_df}{dataframe containing some covariate and response information}

\item{modeling_options}{a named list, additional elements of which
overrides learner-dependent options set within the code}
}
\value{
function that performs forecasting with proper options set
}
\description{
Perform stacked forecasting
}
